+++ Tail/Components/PhysicsComponent.h
@@ -8,7 +8,7 @@ public:
     PhysicsComponent();
     ~PhysicsComponent();
     virtual void Update();
-    virtual void FixedUpdate(sf::RenderWindow& window, double delta, const Level& level);
+    virtual void FixedUpdate(sf::RenderWindow& window, float delta, const Level& level);
     
     sf::Vector2f Velocity() { return m_velocity; }
