+++ Tail/Enemy.h
@@ -10,18 +10,18 @@ class RenderWindow;
 class Enemy : public Entity {
     friend class Breed;
 public:
-    virtual void FixedUpdate(sf::RenderWindow& window, double delta, const Level& level);
+    virtual void FixedUpdate(sf::RenderWindow& window, float delta, const Level& level);
     virtual void HandleCommand(Command command);
 
     const bool IsAlive() const;
-    const double GetHealth() const;
-    const double GetWalkingSpeed() const;
+    const float GetHealth() const;
+    const float GetWalkingSpeed() const;
     const std::string& GetFollowingSprite() const;
     const std::string& GetIdlingSprite() const;
 private:
     Enemy(Breed& breed);
     Breed& m_breed;
-    double m_health;
+    float m_health;
     EnemyState* m_state;
 };
