+++ Tail/IdlingState.h
@@ -9,10 +9,10 @@ class IdlingState : public EnemyState {
 public:
     IdlingState();
     virtual EnemyState* HandleCommand(Command command);
-    virtual void Update(Enemy& enemy, sf::RenderWindow& window, double delta, const Level& level);
+    virtual void Update(Enemy& enemy, sf::RenderWindow& window, float delta, const Level& level);
     virtual void Enter(Enemy& enemy);
 private:
-    double m_idlingTimer;
+    float m_idlingTimer;
 };
 
 #endif
